% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{formatTxSpots}
\alias{formatTxSpots}
\alias{addTxSpots}
\title{Read and process transcript spots geometry for SFE}
\usage{
formatTxSpots(
  file,
  dest = c("rowGeometry", "colGeometry", "imgData"),
  spatialCoordsNames = c("global_x", "global_y", "global_z"),
  gene_col = "gene",
  cell_col = "...1",
  z = 3L,
  phred_col = "qv",
  min_phred = 20,
  split_col = NULL,
  not_in_cell_id = "-1",
  extent = NULL,
  digits = 6L,
  file_out = NULL
)

addTxSpots(
  sfe,
  file,
  sample_id = NULL,
  dest = c("rowGeometry", "colGeometry", "imgData"),
  spatialCoordsNames = c("global_x", "global_y", "global_z"),
  gene_col = "gene",
  cell_col = "...1",
  z = 3L,
  phred_col = "qv",
  min_phred = 20,
  split_col = NULL,
  not_in_cell_id = "-1",
  extent = NULL,
  digits = 6L,
  file_out = NULL
)
}
\arguments{
\item{file}{File with the transcript spot coordinates. Should be one row per
spot when read into R and should have columns for coordinates on each axis,
gene the transcript is assigned to, and optionally cell the transcript is
assigned to. Must be csv, tsv, or parquet.}

\item{dest}{Where in the SFE object to store the spot geometries. This
affects how the data is processed. If "rowGeometry", then all spots for
each gene will be a `MULTIPOINT` geometry, regardless of whether they are
in cells or which cells they are assigned to. If "colGeometry", then spots
for each gene assigned to a cell of interest will be a `MULTIPOINT`
geometry. If "imgData", the spots will be written to a TIFF file as raster
to be read to `imgData(sfe)`.}

\item{spatialCoordsNames}{Column names for the x, y, and optionally z
coordinates of the spots. The defaults are for Vizgen.}

\item{gene_col}{Column name for genes.}

\item{cell_col}{Column name for cell IDs, ignored if `dest = "rowGeometry"`.}

\item{z}{Index of z plane to read. Can be "all" to read all z-planes into
MULTIPOINT geometries with XYZ coordinates. If z values are not integer,
then spots with all z values will be read.}

\item{phred_col}{Column name for Phred scores of the spots.}

\item{min_phred}{Minimum Phred score to keep spot. By default 20, the
conventional threshold indicating "acceptable", meaning that there's 1%
chance that the spot was decoded in error.}

\item{split_col}{Categorical column to split the geometries, such as cell
compartment the spots are assigned to as in the "CellComp" column in CosMX
output.}

\item{not_in_cell_id}{Value of cell ID indicating that the spot is not
assigned to any cell, such as "-1" in Vizgen MERFISH.}

\item{extent}{object of class SpatExtent. If present, the arguments xmin, xmax, ymin and ymax are ignored}

\item{digits}{integer to set the precision for detecting whether points are on a regular grid (a low number of digits is a low precision). Only used when \code{type="xyz"}}

\item{file_out}{Name of file to save the geometry or raster to disk.
Especially when the geometries are so large that it's unwieldy to load
everything into memory. If this file (or directory for multiple files)
already exists, then the existing file(s) will be read, skipping the
processing. When writing the file, extensions supplied are ignored and
extensions are determined based on `dest`.}

\item{sfe}{A `SpatialFeatureExperiment` object.}

\item{sample_id}{A \code{character} sample identifier, which matches the
\code{sample_id} in \code{\link[SpatialExperiment]{imgData}}. The \code{sample_id} will also
be stored in a new column in \code{\link[SpatialExperiment]{colData}}, if not already present.
Default = \code{sample01}.}
}
\value{
A sf data frame for vector geometries if `file_out` is not set.
  `SpatRaster` for raster. If there are multiple files written, such as when
  splitting by cell compartment or when `dest = "colGeometry"`, then a
  directory with the same name as `file_out` will be created (but without the
  extension) and the files are written to that directory with informative
  names. `parquet` files that can be read with `sfarrow::st_read_parquet` is
  written for vector geometries and `geotiff` files that can be read with
  `terra::rast` is written for raster.

The `sf` data frame or the `SpatRaster` (for `dest = "imgData"`).
}
\description{
The function `formatTxSpots` reads the transcript spot coordinates of
smFISH-based data and formats the data. The data is not added to an SFE
object. If the file specified in `file_out` already exists, then this file
will be read instead of the original file in the `file` argument, so the
processing is not run multiple times. The function `addTxSpots` adds the data
read and processed in `formatTxSpots` to the SFE object, and reads all
transcript spot data. To only read a subset of transcript spot data, first
use `formatTxSpots` to write the re-formatted data to disk. Then read the
specific subset and add them separately to the SFE object with the setter
functions.
}
\examples{
# example code

}
