% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotGeometries.R
\name{annotGeometries}
\alias{annotGeometries}
\alias{annotGeometries,SpatialFeatureExperiment-method}
\alias{annotGeometries<-,SpatialFeatureExperiment-method}
\alias{annotGeometryNames,SpatialFeatureExperiment-method}
\alias{annotGeometryNames<-,SpatialFeatureExperiment,character-method}
\alias{annotGeometry,SpatialFeatureExperiment,missing-method}
\alias{annotGeometry,SpatialFeatureExperiment,numeric-method}
\alias{annotGeometry,SpatialFeatureExperiment,character-method}
\alias{annotGeometry<-,SpatialFeatureExperiment,missing-method}
\alias{annotGeometry<-,SpatialFeatureExperiment,numeric-method}
\alias{annotGeometry<-,SpatialFeatureExperiment,character-method}
\title{Annotation geometry methods}
\usage{
\S4method{annotGeometries}{SpatialFeatureExperiment}(x)

\S4method{annotGeometries}{SpatialFeatureExperiment}(x, ...) <- value

\S4method{annotGeometryNames}{SpatialFeatureExperiment}(x)

\S4method{annotGeometryNames}{SpatialFeatureExperiment,character}(x) <- value

\S4method{annotGeometry}{SpatialFeatureExperiment,missing}(x, type)

\S4method{annotGeometry}{SpatialFeatureExperiment,numeric}(x, type)

\S4method{annotGeometry}{SpatialFeatureExperiment,character}(x, type)

\S4method{annotGeometry}{SpatialFeatureExperiment,missing}(x, type) <- value

\S4method{annotGeometry}{SpatialFeatureExperiment,numeric}(x, type, ...) <- value

\S4method{annotGeometry}{SpatialFeatureExperiment,character}(x, type, ...) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{...}{\code{spatialCoordsNames, spotDiameter, geometryType} passed to
\code{\link{df2sf}}. For \code{dimGeometries<-} only: \code{geometryType}
can be a character vector of the geometry type of each data frame in the
list of the same length as the list if the data frames specify different
types of geometries.}

\item{value}{Value to set. For \code{annotGeometry}, must be a \code{sf} data
frame, or an ordinary data frame that can be converted to a \code{sf} data
frame (see \code{\link{df2sf}}). For \code{annotGeometries}, must be a list
of such \code{sf} or ordinary data frames.}
}
\description{
"Annotation geometry" refers to Simple Feature (\code{sf}) geometries NOT
associated with rows (features, genes) or columns (cells or spots) of the
gene count matrix in the \code{SpatialFeatureExperiment} object. So there can
be any number of rows in the \code{sf} data frame specifying the geometry.
Examples of such geometries are tissue boundaries, pathologist annotation of
histological regions, and objects not characterized by columns of the gene
count matrix (e.g. nuclei segmentation in a Visium dataset where the columns
are Visium spots). This page documents getters and setters for the annotation
geometries. Internally, annotation geometries are stored in
\code{int_metadata}.
}
