% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{bbox,SpatialFeatureExperiment-method}
\alias{bbox,SpatialFeatureExperiment-method}
\alias{bbox}
\title{Find bounding box of SFE objects}
\usage{
\S4method{bbox}{SpatialFeatureExperiment}(sfe, sample_id)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{sample_id}{Sample(s) whose bounding box(es) to find. The bounding box
would be for the union of all \code{colGeometries} and
\code{annotGeometries} associated with each sample.}
}
\value{
For one sample, then a named vector with names \code{xmin},
  \code{ymin}, \code{xmax}, and \code{ymax} specifying the bounding box. For
  multiple samples, then a matrix whose columns are samples and whose rows
  delineate the bounding box.
}
\description{
Find bounding box of the union of all \code{colGeometries} and
\code{annotGeometries} of each sample in the SFE object. This can be used to
remove empty space so the tissue and geometries have one corner at the origin
so all samples will be on comparable coordinates.
}
