% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox_center}
\alias{bbox_center}
\title{Find center of bounding box}
\usage{
bbox_center(bbox)
}
\arguments{
\item{bbox}{A numeric vector of length 4 with names xmin, xmax, ymin, ymax,
in any order.}
}
\value{
A numeric vector of length 2.
}
\description{
Get x-y coordinates of the center of any bounding box
}
\examples{
bbox <- c(xmin = 0, xmax = 100, ymin = 0, ymax = 80)
bbox_center(bbox)
}
\concept{Utilities}
