% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{BioFormatsImage}
\alias{BioFormatsImage}
\title{On disk representation of BioFormats images in SFE object}
\usage{
BioFormatsImage(path, ext = NULL, isFull = TRUE)
}
\arguments{
\item{path}{Path to an OME-TIFF image file.}

\item{ext}{Numeric vector with names "xmin", "xmax", "ymin", "ymax" in microns
indicating the spatial extent covered by the image. If \code{NULL}, then the
extent will be inferred from the metadata, from physical pixel size and the
number of pixels.}

\item{isFull}{Logical, if the extent specified in \code{ext} is the full
extent. If \code{ext = NULL} so it will be inferred from metadata then
\code{isFull = TRUE} will be set internally.}
}
\value{
A \code{BioFormatsImage} object.
}
\description{
At present, the \code{BioFormatsImage} is designed for OME-TIFF from Xenium
and has not been tested on other formats that can be read with
\code{BioFormats}. The image is not loaded into memory, and when it is, the
the \code{BioFormatsImage} object is converted into \code{\link{EBImage}}
because the loaded image is of a class that inherits from
\code{\link{EBImage::Image}}. The \code{\link{EBImage}} class is a thin
wrapper inheriting from \code{VirtualSpatialImage} so it's compatible with
\code{SpatialExperiment} from which SFE is derived.
}
\seealso{
[isFull()]
}
\concept{Image and raster}
