% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{callMeta}
\alias{callMeta}
\title{Call metadata of an object}
\usage{
callMeta(object = NULL)
}
\arguments{
\item{object}{A \code{SpatialFeatureExperiment}, \code{SpatialExperiment}, or \code{SummarizedExperiment} object.}
}
\value{
A \code{data.frame} object
}
\description{
Call metadata of an object
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
sfe |> callMeta() |> str()
}
