% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimGeometries.R
\name{dimGeometries}
\alias{dimGeometries}
\alias{dimGeometry}
\alias{dimGeometryNames}
\alias{colGeometry}
\alias{rowGeometry}
\alias{colGeometries}
\alias{rowGeometries}
\alias{colGeometryNames}
\alias{rowGeometryNames}
\alias{colGeometry<-}
\alias{rowGeometry<-}
\alias{colGeometries<-}
\alias{rowGeometries<-}
\alias{colGeometryNames<-}
\alias{rowGeometryNames<-}
\alias{dimGeometry,SpatialFeatureExperiment,missing-method}
\alias{dimGeometry,SpatialFeatureExperiment,numeric-method}
\alias{dimGeometry,SpatialFeatureExperiment,character-method}
\alias{dimGeometries,SpatialFeatureExperiment-method}
\alias{dimGeometry<-}
\alias{dimGeometries<-}
\alias{dimGeometryNames<-}
\alias{dimGeometry<-,SpatialFeatureExperiment,missing-method}
\alias{dimGeometry<-,SpatialFeatureExperiment,numeric-method}
\alias{dimGeometry<-,SpatialFeatureExperiment,character-method}
\alias{dimGeometries<-,SpatialFeatureExperiment-method}
\alias{dimGeometryNames<-,SpatialFeatureExperiment,numeric,character-method}
\alias{dimGeometryNames,SpatialFeatureExperiment-method}
\title{Dimension geometry methods}
\usage{
\S4method{dimGeometries}{SpatialFeatureExperiment}(x, MARGIN = 2, withDimnames = TRUE)

\S4method{dimGeometries}{SpatialFeatureExperiment}(x, MARGIN, withDimnames = TRUE, ...) <- value

\S4method{dimGeometryNames}{SpatialFeatureExperiment}(x, MARGIN)

\S4method{dimGeometryNames}{SpatialFeatureExperiment,numeric,character}(x, MARGIN) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,missing}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,numeric}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,character}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,missing}(x, type, MARGIN, withDimnames = TRUE) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,numeric}(x, type, MARGIN, withDimnames = TRUE, ...) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,character}(x, type, MARGIN, withDimnames = TRUE, ...) <- value

colGeometry(x, type, withDimnames = TRUE)

colGeometry(x, type, withDimnames = TRUE) <- value

colGeometries(x, withDimnames = TRUE)

colGeometries(x, withDimnames = TRUE) <- value

colGeometryNames(x)

colGeometryNames(x) <- value

rowGeometry(x, type, withDimnames = TRUE)

rowGeometry(x, type, withDimnames = TRUE) <- value

rowGeometries(x, withDimnames = TRUE)

rowGeometries(x, withDimnames = TRUE) <- value

rowGeometryNames(x)

rowGeometryNames(x) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{MARGIN}{As in \code{\link{apply}}. 1 stands for rows and 2 stands for
columns.}

\item{withDimnames}{Logical. If \code{TRUE}, then the dimnames (colnames or
rownames) of the gene count matrix should correspond to row names of the
\code{sf} data frames of interest.}

\item{...}{\code{spatialCoordsNames, spotDiameter, geometryType} passed to
\code{\link{df2sf}}. For \code{dimGeometries<-} only: \code{geometryType}
can be a character vector of the geometry type of each data frame in the
list of the same length as the list if the data frames specify different
types of geometries.}

\item{value}{Value to set. For \code{dimGeometry}, must be a \code{sf} data
frame with the same number of rows as size in the dimension of interest, or
an ordinary data frame that can be converted to such a \code{sf} data frame
(see \code{\link{df2sf}}). For \code{dimGeometries}, must be a list of such
\code{sf} or ordinary data frames.}
}
\description{
"Dimension geometry" refers to Simple Feature (\code{sf}) geometries
associated with rows (features, genes) or columns (cells or spots) of the
gene count matrix in the \code{SpatialFeatureExperiment} object. For each
dimension, the number of rows in the \code{sf} data frame specifying the
geometries must match the size of the dimension of interest. For example,
there must be the same number of rows in the \code{sf} data frame describing
cells as there are cells in the gene count matrix. This page documents
getters and setters for the dimension geometries. The getters and setters are
implemented in a way similar to those of \code{reducedDims} in
\code{SingleCellExperiment}.
}
