% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.check_st_valid}
\alias{.check_st_valid}
\title{sanity on geometries to remove any self-intersection}
\usage{
.check_st_valid(sf_df = NULL, dist = 0)
}
\description{
sanity on geometries to remove any self-intersection
}
