% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imageIDs}
\alias{imageIDs}
\title{Show all image_ids in the SFE object}
\usage{
imageIDs(sfe)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A character vector of \code{image_ids}.
}
\description{
The title is self-explanatory. Some functions require \code{image_id} to get
or set images.
}
\examples{
fp <- system.file(file.path("extdata", "sample01"),
package = "SpatialFeatureExperiment")
sfe <- read10xVisiumSFE(fp, type = "sparse")
imageIDs(sfe)
}
\concept{Utilities}
