% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{isFull,BioFormatsImage-method}
\alias{isFull,BioFormatsImage-method}
\title{Check if the extent is the full extent of the image}
\usage{
\S4method{isFull}{BioFormatsImage}(x)
}
\arguments{
\item{x}{A \code{\link{BioFormatsImage}} object.}
}
\value{
Logical scalar indicating whether the extent is the full extent.
}
\description{
Check if the extent is the full extent of the image
}
