% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE-class.R
\name{make_empty_geometry}
\alias{make_empty_geometry}
\title{Make empty geometry}
\usage{
make_empty_geometry()
}
\value{
An \code{sf} data frame whose geometry is an empty GEOMETRYCOLLECTION.
}
\description{
For default arguments in some functions with the right type.
}
