% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readVisiumHDCellSeg}
\alias{readVisiumHDCellSeg}
\title{Reader for Visium HD cell segmented output.}
\usage{
readVisiumHDCellSeg(td, res)
}
\arguments{
\item{td}{path to unzipped Visium HD data download with `/segmented_outputs`}

\item{res}{"hires" or "lowres" for what .png to read in.}
}
\value{
a `SpatialFeatureExperiment` object.
}
\description{
Reader for Visium HD cell segmented output.
}
\examples{
vhdsfe <- readVisiumHDCellSeg(td = "~/Desktop", res = "hires")
}
