% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{colData}
\alias{rowData}
\alias{colData<-}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{spatialCoordsNames}
\alias{getImg}
\alias{imgData}
\alias{rmvImg}
\alias{counts}
\alias{logcounts}
\alias{reducedDim}
\title{Functions re-exported from other packages}
\usage{
colData(x, ...)

rowData(x, use.names = TRUE, ...)

colData(x, ...) <- value

spatialCoords(x, ...)

spatialCoords(x) <- value

spatialCoordsNames(x)

getImg(x, ...)

imgData(x)

rmvImg(x, ...)

counts(object, ...)

logcounts(object, ...)

reducedDim(x, type, ...)
}
\description{
These are some commonly used getters and setters of classes that SFE inherits
so you don't have to separately attach those packages to use these functions.
}
