% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{removeEmptySpace}
\alias{removeEmptySpace}
\title{Remove empty space}
\usage{
removeEmptySpace(sfe, sample_id = "all")
}
\arguments{
\item{sample_id}{Samples to crop. Can be multiple samples, or "all", which
means all samples. For multiple samples, \code{y} may have column
\code{sample_id} indicating which geometry subsets which sample. Only
samples included in the \code{sample_id} column are subsetted. If there is
no \code{sample_id} column or \code{y} is not specified, then the same
geometry or bounding box is used to subset all samples specified in the
\code{sample_id} argument.}
}
\value{
An SFE object with empty space removed.
}
\description{
For each sample independently, anything outside the bounding box of all
colGeometries and annotGeometries is removed, and the coordinates are changed
to within the bounding box. So when using geom_sf for plotting, different
samples will have more comparable coordinates, as free scales can't be used
when facetting geom_sf.
}
\note{
Unlike other functions in this package, this function operates on all
samples by default.
}
