% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleIDs}
\alias{sampleIDs}
\title{Get all unique sample IDs}
\usage{
sampleIDs(sfe)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A character vector of all unique entries of the \code{sample_id}
column in \code{colData(x)}.
}
\description{
The title is self-explanatory.
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
sampleIDs(sfe)
}
\concept{Utilities}
