% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE-class.R
\name{show,SpatialFeatureExperiment-method}
\alias{show,SpatialFeatureExperiment-method}
\title{Print method for SpatialFeatureExperiment}
\usage{
\S4method{show}{SpatialFeatureExperiment}(object)
}
\arguments{
\item{object}{A \code{SpatialFeatureExperiment} object.}
}
\value{
None (invisible \code{NULL}).
}
\description{
Printing summaries of \code{colGeometries}, \code{rowGeometries}, and
\code{annotGeometries} in addition to what's shown for
\code{SpatialExperiment}. Geometry names and types are printed.
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
sfe # The show method is implicitly called
}
\concept{SpatialFeatureExperiment class}
