% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{SpatRasterImage}
\alias{SpatRasterImage}
\alias{SpatRasterImage-class}
\title{SpatRaster representation of images in SFE objects}
\usage{
SpatRasterImage(img)
}
\arguments{
\item{img}{A \code{\link{SpatRaster}} or \code{\link{PackedSpatRaster}} object.}
}
\value{
Methods for \code{SpatRasterImage} return a modified
  \code{SpatRasterImage}, and methods for SFE return a modified SFE object.
}
\description{
\code{SpatialFeatureExperiment} and the \code{Voyager} package work with
images differently from \code{SpatialExperiment}. In SFE and
\code{Voyager}'s, plotting functions for SFE objects, the images can be read
with \code{\link{rast}} and represented as \code{SpatRaster}, so the image is
not entirely loaded into memory unless necessary. Plotting will not load a
large image into memory; rather the image will be downsampled and the
downsampled version is plotted.
}
\note{
If the image is already a GeoTIFF file that already has an extent, then
  the extent associated with the file will be honored and the \code{extent}
  and \code{scale_fct} arguments are ignored. Also, when the image is
  transposed, it is flipped about the axis going from top left to bottom
  right.
}
\examples{
# Example code
}
\concept{Image and raster}
