% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitComponent}
\alias{splitComponent}
\alias{splitComponent,SpatialFeatureExperiment-method}
\alias{splitComponent,sf-method}
\alias{splitComponent,sfc-method}
\title{Split by graph component}
\usage{
\S4method{splitComponent}{SpatialFeatureExperiment}(x, colGraphName = 1L, min_cells = 100)

\S4method{splitComponent}{sf}(
  x,
  min_cells = 100,
  distance_cutoff = 50,
  BNPARAM = NULL,
  BPPARAM = SerialParam()
)

\S4method{splitComponent}{sfc}(
  x,
  min_cells = 100,
  distance_cutoff = 50,
  BNPARAM = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{A SFE object, or \code{sf} or \code{sfc}.}

\item{colGraphName}{Name of graph to use for \code{splitComponent}. We
recommend distance based neighbors (`dnearneigh` in
\code{\link{findSpatialNeighbors}}), and recommend NOT using k nearest
neighbors (`knearneigh`) or triangulation (`tri2nb`).}

\item{min_cells}{Minimum number of cells per graph component; components with
fewer than this number of cells are considered debris and removed.}

\item{distance_cutoff}{For the \code{sf} and \code{sfc} methods, the distance
cutoff used to construct a distance-based spatial neighborhood graph, in
the same unit as the spatial coordinates. Anything within this distance is
considered a neighbor.}

\item{BNPARAM}{A \code{\link[BiocNeighbors]{BiocNeighborParam}} object specifying the
algorithm to find k nearest neighbors and distance based neighbors with
\code{nn_method = "bioc"}. For distance based neighbors, only
\code{\link[BiocNeighbors]{KmknnParam}} and \code{\link[BiocNeighbors]{VptreeParam}} are applicable.}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} object for multithreading.
Only used for k nearest neighbor and distance based neighbor with
\code{nn_method = "bioc"}.}
}
\value{
A list of the same object type as the input, each element for one
  component.
}
\description{
Split either a SFE object or \code{sf} or \code{sfc} by graph component of a
spatial neighborhood graph.
}
