% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{toEBImage}
\alias{toEBImage}
\alias{toEBImage,SpatRasterImage-method}
\title{Convert images to EBImage}
\usage{
\S4method{toEBImage}{BioFormatsImage}(x, resolution = 4L)

\S4method{toEBImage}{SpatRasterImage}(x, maxcell = 1e+07)
}
\arguments{
\item{x}{Either a \code{BioFormatsImage} or \code{SpatRasterImage} object.}

\item{resolution}{Integer, which resolution in the \code{BioFormatsImage} to
read and convert. Defaults to 4, which is a lower resolution. Ignored if only
1 resolution is present.}
}
\value{
A \code{EBImage} object. The image is loaded into memory.
}
\description{
The \code{EBImage} class is a thin wrapper around the \code{Image} class in
\code{EBImage} so it inherits from \code{VirtualSpatialImage} as required by
\code{SpatialExperiment} and has extent as used in Voyager's plotting
functions. This function converts \code{SpatRasterImage} (thin wrapper around
the class in \code{terra}) and \code{BioFormatsImage} into \code{EBImage} for
image operations as implemented in the \code{EBImage} package.
}
\seealso{
toSpatRasterImage
}
\concept{Image and raster}
