% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE-class.R
\name{unit,SpatialFeatureExperiment-method}
\alias{unit,SpatialFeatureExperiment-method}
\alias{unit}
\title{Get unit of a SpatialFeatureExperiment}
\usage{
\S4method{unit}{SpatialFeatureExperiment}(x)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A string for the name of the unit. At present it's merely a
  string and \code{udunits} is not used.
}
\description{
Length units can be microns or pixels in full resolution image in SFE
objects.
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
SpatialFeatureExperiment::unit(sfe)
}
\concept{SpatialFeatureExperiment class}
