% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visium_row_col}
\alias{visium_row_col}
\title{Row and columns of Visium barcodes on the slide}
\format{
A data frame with 4992 rows with columns barcode, col, and row.
}
\source{
Space Ranger 1.3.1
}
\usage{
data(visium_row_col)
}
\description{
From Space Ranger 1.3.1.
}
\concept{Utilities}
\keyword{datasets}
