% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimGeometries.R
\name{addVisiumSpotPoly}
\alias{addVisiumSpotPoly}
\title{Add Visium spot polygons to colGeometry}
\usage{
addVisiumSpotPoly(x, spotDiameter)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{spotDiameter}{Spot diameter for technologies with arrays of spots of
fixed diameter per slide, such as Visium, ST, DBiT-seq, and slide-seq. The
diameter must be in the same unit as the coordinates in the *Geometry
arguments. Ignored for geometries that are not POINT or MULTIPOINT.}
}
\value{
A SFE object with a new colGeometry called spotPoly, which has
polygons of the spots.
}
\description{
For adding the spot polygons to SFE objects converted from SPE.
}
