% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateTx}
\alias{aggregateTx}
\alias{aggregateTxTech}
\title{Aggregate transcript spots from file}
\usage{
aggregateTx(
  file,
  df = NULL,
  by = NULL,
  spatialCoordsNames = c("X", "Y", "Z"),
  gene_col = "gene",
  phred_col = "qv",
  min_phred = 20,
  flip = FALSE,
  cellsize = NULL,
  square = TRUE,
  flat_topped = FALSE,
  new_geometry_name = "bins",
  unit = "micron"
)

aggregateTxTech(
  data_dir,
  df = NULL,
  by = NULL,
  tech = c("Vizgen", "Xenium", "CosMX"),
  min_phred = 20,
  flip = FALSE,
  cellsize = NULL,
  square = TRUE,
  flat_topped = FALSE,
  new_geometry_name = "bins"
)
}
\arguments{
\item{file}{File with the transcript spot coordinates. Should be one row per
spot when read into R and should have columns for coordinates on each axis,
gene the transcript is assigned to, and optionally cell the transcript is
assigned to. Must be csv, tsv, or parquet.}

\item{df}{If the file is already loaded into memory, a data frame (sf) with
columns for the x, y, and optionally z coordinates and gene assignment of
each transcript spot. If specified, then argument \code{file} will be
ignored.}

\item{by}{A \code{sfc} or \code{sf} object for spatial aggregation.}

\item{spatialCoordsNames}{Column names for the x, y, and optionally z
coordinates of the spots. The defaults are for Vizgen.}

\item{gene_col}{Column name for genes.}

\item{phred_col}{Column name for Phred scores of the spots.}

\item{min_phred}{Minimum Phred score to keep spot. By default 20, the
conventional threshold indicating "acceptable", meaning that there's 1%
chance that the spot was decoded in error.}

\item{flip}{Logical, whether to flip the geometry to match image. Here the y
coordinates are simply set to -y, so the original bounding box is not
preserved. This is consistent with \code{readVizgen} and \code{readXenium}.}

\item{data_dir}{Top level output directory.}

\item{tech}{Which technology whose output to read, must be one of "Vizgen",
"Xenium", or "CosMX" though more technologies may be added later.}
}
\value{
A SFE object with count matrix for number of spots of each gene in
  each geometry. Geometries with no spot are removed.
}
\description{
This function reads the transcript spot file from the standard output of the
commercial technologies (not GeoParquet) for spatial aggregation where the
spots are assigned to polygons such as cells or spatial bins. Presets for
Xenium, MERFISH, and CosMX are available.
}
\note{
The resulting SFE object often includes geometries (e.g. grid cells)
  outside tissue, because there can be transcript spots detected outside the
  tissue.
}
