% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop an SFE object with a geometry}
\usage{
crop(
  x,
  y = NULL,
  colGeometryName = 1L,
  pred = st_intersects,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{x}{An SFE object.}

\item{y}{An object of class \code{sf}, \code{sfc}, or \code{sfg} with which
to crop the SFE object. Optional if \code{xmin}, \code{xmax}, \code{ymin},
and \code{ymax} are specified for a bounding box.}

\item{colGeometryName}{Column geometry to used to indicate which cells/spots
to keep.}

\item{pred}{A geometric binary predicate function to indicate which
cells/spots to keep, defaults to \code{\link{st_intersects}}.}

\item{xmin}{Minimum x coordinate of bounding box. Ignored if \code{y} is
specified.}

\item{xmax}{Maximum x coordinate of bounding box.}

\item{ymin}{Minimum y coordinate of bounding box.}

\item{ymax}{Maximum y coordinate of bounding box.}
}
\value{
An SFE object.
}
\description{
Returns an SFE object whose specified \code{colGeometry} returns \code{TRUE}
with a geometric predicate function (usually intersects) with another
geometry of interest. This can be used to subset an SFE object with a tissue
boundary or histological region polygon, or crop away empty spaces.
}
