% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2sf.R
\name{df2sf}
\alias{df2sf}
\title{From ordinary data frame to sf to construct SFE object}
\usage{
df2sf(
  df,
  spatialCoordsNames = c("x", "y"),
  spotDiameter = NA,
  geometryType = c("POINT", "LINESTRING", "POLYGON", "MULTIPOINT", "MULTILINESTRING",
    "MULTIPOLYGON")
)
}
\arguments{
\item{df}{An ordinary data frame, i.e. not \code{sf}. Or a matrix that can be
converted to a data frame.}

\item{spatialCoordsNames}{Column names in \code{df} that specify spatial
coordinates.}

\item{spotDiameter}{Spot diameter for technologies with arrays of spots of
fixed diameter per slide, such as Visium, ST, DBiT-seq, and slide-seq. The
diameter must be in the same unit as the coordinates in the *Geometry
arguments. Ignored for geometries that are not POINT or MULTIPOINT.}

\item{geometryType}{Type of geometry to convert the ordinary data frame to.
If the geometry in \code{df} is de facto points, then this argument will be
ignored and the returned \code{sf} will have geometry type POINT.}
}
\value{
An \code{sf} object.
}
\description{
While the \code{SpatialFeatureExperiment} constructor and \code{*Geometry}
replacement methods can convert properly formatted ordinary data frames into
\code{sf} objects which are used to store the geometries internally, the user
might want to do the conversion, check if the geometry is valid, and inspect
and fix any invalid geometries.
}
