% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{dim,ExtImage-method}
\alias{dim,ExtImage-method}
\title{Find dimensions of ExtImage}
\usage{
\S4method{dim}{ExtImage}(x)
}
\arguments{
\item{x}{A \code{\link{ExtImage}} object.}
}
\value{
An integer vector. As in \code{EBImage}, the first element indicates
number of pixels in the x direction, or number of columns in the image, and
the second element indicates the number of pixels in the y direction. This is
unlike array indexing.
}
\description{
This method exists to make the output of \code{dim()} for \code{ExtImage}
consistent with that of \code{Image} which \code{ExtImage} inherits from,
overriding the \code{VirtualSpatialImage} method.
}
\seealso{
Other image methods: 
\code{\link{SFE-image}},
\code{\link{affineImg}()},
\code{\link{cropImg}()},
\code{\link{dim,BioFormatsImage-method}},
\code{\link{ext}()},
\code{\link{imgRaster}()},
\code{\link{imgSource}()},
\code{\link{mirrorImg}()},
\code{\link{rotateImg}()},
\code{\link{scaleImg}()},
\code{\link{translateImg}()},
\code{\link{transposeImg}()}
}
\concept{Image methods}
\concept{image methods}
