% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimGeometries.R
\name{dimGeometries}
\alias{dimGeometries}
\alias{dimGeometry}
\alias{dimGeometryNames}
\alias{colGeometry}
\alias{rowGeometry}
\alias{colGeometries}
\alias{rowGeometries}
\alias{colGeometryNames}
\alias{rowGeometryNames}
\alias{colGeometry<-}
\alias{rowGeometry<-}
\alias{colGeometries<-}
\alias{rowGeometries<-}
\alias{colGeometryNames<-}
\alias{rowGeometryNames<-}
\alias{dimGeometry,SpatialFeatureExperiment,missing-method}
\alias{dimGeometry,SpatialFeatureExperiment,numeric-method}
\alias{dimGeometry,SpatialFeatureExperiment,character-method}
\alias{dimGeometries,SpatialFeatureExperiment-method}
\alias{dimGeometry<-}
\alias{dimGeometries<-}
\alias{dimGeometryNames<-}
\alias{dimGeometry<-,SpatialFeatureExperiment,missing-method}
\alias{dimGeometry<-,SpatialFeatureExperiment,numeric-method}
\alias{dimGeometry<-,SpatialFeatureExperiment,character-method}
\alias{dimGeometries<-,SpatialFeatureExperiment-method}
\alias{dimGeometryNames<-,SpatialFeatureExperiment,numeric,character-method}
\alias{dimGeometryNames,SpatialFeatureExperiment-method}
\alias{spotPoly}
\alias{spotPoly<-}
\alias{ROIPoly}
\alias{ROIPoly<-}
\alias{cellSeg}
\alias{cellSeg<-}
\alias{nucSeg}
\alias{nucSeg<-}
\alias{txSpots}
\alias{txSpots<-}
\title{Dimension geometry methods}
\usage{
\S4method{dimGeometries}{SpatialFeatureExperiment}(x, MARGIN = 2, withDimnames = TRUE)

\S4method{dimGeometries}{SpatialFeatureExperiment}(x, MARGIN, withDimnames = TRUE, ...) <- value

\S4method{dimGeometryNames}{SpatialFeatureExperiment}(x, MARGIN)

\S4method{dimGeometryNames}{SpatialFeatureExperiment,numeric,character}(x, MARGIN) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,missing}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,numeric}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,character}(x, type, MARGIN, withDimnames = TRUE)

\S4method{dimGeometry}{SpatialFeatureExperiment,missing}(x, type, MARGIN, withDimnames = TRUE) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,numeric}(x, type, MARGIN, withDimnames = TRUE, ...) <- value

\S4method{dimGeometry}{SpatialFeatureExperiment,character}(x, type, MARGIN, withDimnames = TRUE, ...) <- value

colGeometry(x, type, withDimnames = TRUE)

colGeometry(x, type, withDimnames = TRUE) <- value

colGeometries(x, withDimnames = TRUE)

colGeometries(x, withDimnames = TRUE) <- value

colGeometryNames(x)

colGeometryNames(x) <- value

rowGeometry(x, type, withDimnames = TRUE)

rowGeometry(x, type, withDimnames = TRUE) <- value

rowGeometries(x, withDimnames = TRUE)

rowGeometries(x, withDimnames = TRUE) <- value

rowGeometryNames(x)

rowGeometryNames(x) <- value

spotPoly(x, withDimnames = TRUE)

spotPoly(x, withDimnames = TRUE) <- value

ROIPoly(x, withDimnames = TRUE)

ROIPoly(x, withDimnames = TRUE) <- value

cellSeg(x, withDimnames = TRUE)

cellSeg(x, withDimnames = TRUE) <- value

nucSeg(x, withDimnames = TRUE)

nucSeg(x, withDimnames = TRUE) <- value

txSpots(x, withDimnames = TRUE)

txSpots(x, withDimnames = TRUE) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{MARGIN}{As in \code{\link{apply}}. 1 stands for rows and 2 stands for
columns.}

\item{withDimnames}{Logical. If \code{TRUE}, then the dimnames (colnames or
rownames) of the gene count matrix should correspond to row names of the
\code{sf} data frames of interest.}

\item{...}{\code{spatialCoordsNames, spotDiameter, geometryType} passed to
\code{\link{df2sf}}. Defaults are the same as in \code{\link{df2sf}}. For
\code{dimGeometries<-} only: \code{geometryType} can be a character vector
of the geometry type of each data frame in the list of the same length as
the list if the data frames specify different types of geometries.}

\item{value}{Value to set. For \code{dimGeometry}, must be a \code{sf} data
frame with the same number of rows as size in the dimension of interest, or
an ordinary data frame that can be converted to such a \code{sf} data frame
(see \code{\link{df2sf}}). For \code{dimGeometries}, must be a list of such
\code{sf} or ordinary data frames.}
}
\description{
"Dimension geometry" refers to Simple Feature (\code{sf}) geometries
associated with rows (features, genes) or columns (cells or spots) of the
gene count matrix in the \code{SpatialFeatureExperiment} object. For each
dimension, the number of rows in the \code{sf} data frame specifying the
geometries must match the size of the dimension of interest. For example,
there must be the same number of rows in the \code{sf} data frame describing
cells as there are cells in the gene count matrix. This page documents
getters and setters for the dimension geometries. The getters and setters are
implemented in a way similar to those of \code{reducedDims} in
\code{SingleCellExperiment}.
}
\details{
These are convenience wrappers for getters and setters of special geometries:
\describe{ \item{colGeometry/ies}{dimGeometry/ies with MARGIN = 2, for
geometries associated with columns of the gene count matrix (cells/Visium
spots/samples).} \item{rowGeometry/ies}{dimGeometry/ies with MARGIN = 1, for
geometries associated with rows of the gene count matrix (genes/features).}
\item{spotPoly}{Polygons of spots from technologies such as Visium, ST, and
slide-seq, which do not correspond to cells. Centroids of the polygons are
stored in \code{spatialCoords} of the underlying \code{SpatialExperiment}
object.} \item{ROIPoly}{Polygons of regions of interest (ROIs) from
technologies such as laser capture microdissection (LCM) and GeoMX DSP. These
should correspond to columns of the gene count matrix.} \item{cellSeg}{Cell
segmentation polygons. If the columns of the gene count matrix are single
cells, then this is stored in \code{colGeometries}. Otherwise, this is stored
in \code{\link{annotGeometries}}.} \item{nucSeg}{Similar to \code{cellSeg},
but for nuclei rather than whole cell.} \item{txSpots}{POINT or MULTIPOINT
geometries of transcript spots of single molecular resolution technologies,
stored in \code{rowGeometries}.} }
}
