% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{imgSource}
\alias{imgSource}
\alias{imgSource,SpatRasterImage-method}
\alias{imgSource,BioFormatsImage-method}
\alias{imgSource,ExtImage-method}
\title{Source of images that are on disk}
\usage{
\S4method{imgSource}{SpatRasterImage}(x)

\S4method{imgSource}{BioFormatsImage}(x)

\S4method{imgSource}{ExtImage}(x)
}
\arguments{
\item{x}{An object of class \code{*Image} as implemented in this package.}
}
\value{
String, file path to the original image on disk. For
  \code{SpatRasterImage}, if the image is loaded into memory, then
  \code{NULL}.
}
\description{
Get the file path of images that are on disk and not read into memory. Only
applies to \code{SpatRasterImage} and \code{BioFormatsImage}.
}
\seealso{
Other image methods: 
\code{\link{SFE-image}},
\code{\link{affineImg}()},
\code{\link{cropImg}()},
\code{\link{dim,BioFormatsImage-method}},
\code{\link{ext}()},
\code{\link{imgRaster}()},
\code{\link{mirrorImg}()},
\code{\link{rotateImg}()},
\code{\link{scaleImg}()},
\code{\link{translateImg}()},
\code{\link{transposeImg}()}
}
\concept{Image methods}
\concept{image methods}
