% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-Voyager.R
\name{internal-Voyager}
\alias{internal-Voyager}
\alias{.value2df}
\alias{.check_features}
\alias{.warn_symbol_duplicate}
\alias{.symbol2id}
\alias{.check_sample_id}
\alias{.rm_empty_geometries}
\title{Internal functions also used in Voyager}
\usage{
.value2df(value, use_geometry, feature = NULL)

.check_features(x, features, colGeometryName = NULL)

.warn_symbol_duplicate(x, symbols)

.symbol2id(x, features)

.check_sample_id(x, sample_id, one = TRUE)

.rm_empty_geometries(g, MARGIN)
}
\value{
Internal
}
\description{
Not meant for the user, but exporting to be used internally in Voyager. But
one day I may clean these up and remove the internal note for people building
on top of SFE.
}
\keyword{internal}
