% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localResults.R
\docType{methods}
\name{localResults}
\alias{localResults}
\alias{localResults<-}
\alias{localResult}
\alias{localResult<-}
\alias{localResultNames}
\alias{localResultNames<-}
\alias{localResults,SpatialFeatureExperiment-method}
\alias{localResults<-,SpatialFeatureExperiment-method}
\alias{localResultNames,SpatialFeatureExperiment-method}
\alias{localResultNames<-,SpatialFeatureExperiment,character-method}
\alias{localResult,SpatialFeatureExperiment,missing-method}
\alias{localResult,SpatialFeatureExperiment,numeric-method}
\alias{localResult,SpatialFeatureExperiment,character-method}
\alias{localResult<-,SpatialFeatureExperiment,missing-method}
\alias{localResult<-,SpatialFeatureExperiment,numeric-method}
\alias{localResult<-,SpatialFeatureExperiment,character-method}
\title{Organize results from local spatial statistics}
\usage{
\S4method{localResults}{SpatialFeatureExperiment}(x, withDimnames = TRUE, ...)

\S4method{localResults}{SpatialFeatureExperiment}(x, withDimnames = TRUE, ...) <- value

\S4method{localResultNames}{SpatialFeatureExperiment}(x)

\S4method{localResultNames}{SpatialFeatureExperiment,character}(x) <- value

\S4method{localResult}{SpatialFeatureExperiment,missing}(x, type, sample_id = NULL, withDimnames = TRUE)

\S4method{localResult}{SpatialFeatureExperiment,numeric}(x, type, sample_id = NULL, withDimnames = TRUE)

\S4method{localResult}{SpatialFeatureExperiment,character}(x, type, sample_id = NULL, withDimnames = TRUE)

\S4method{localResult}{SpatialFeatureExperiment,missing}(x, type, sample_id = NULL, withDimnames = TRUE) <- value

\S4method{localResult}{SpatialFeatureExperiment,numeric}(x, type, sample_id = NULL, withDimnames = TRUE) <- value

\S4method{localResult}{SpatialFeatureExperiment,character}(x, type, sample_id = NULL, withDimnames = TRUE) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{withDimnames}{Logical. If \code{TRUE}, then the dimnames (colnames or
rownames) of the gene count matrix should correspond to row names of the
\code{sf} data frames of interest.}

\item{...}{Ignored}

\item{value}{Values to set, should be either a matrix or a data frame.}

\item{type}{An integer specifying the index or string specifying the name of
the *Geometry to query or replace. If missing, then the first item in the
*Geometries will be returned or replaced.}

\item{sample_id}{Sample ID to get or set geometries.}
}
\value{
\code{localResults} returns a named list each element of which is a
  set of local results of interest. \code{localResult} returns a matrix or a
  data frame, whichever the original is when it's set.
  \code{localResultNames} returns a character vector. Setters return an SFE
  object with the desired field set.
}
\description{
Local spatial statics like local Moran's I, local Geary's C, Getis-Ord Gi*,
and geographically weighted PCA loadings return values at each spatial
location. Just like dimension reductions, these results are clearly
associated with the braoder SFE object, so they should have a place within
the object. The \code{localResults} field in the SFE object stores these
results that has a value for each spatial location.
}
