% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localResults.R
\docType{methods}
\name{localResults}
\alias{localResults}
\alias{localResults<-}
\alias{localResult}
\alias{localResult<-}
\alias{localResultNames}
\alias{localResultNames<-}
\alias{localResults,SpatialFeatureExperiment,missing,missing-method}
\alias{localResults<-,SpatialFeatureExperiment,missing,missing-method}
\alias{localResults,SpatialFeatureExperiment,ANY,character-method}
\alias{localResults<-,SpatialFeatureExperiment,ANY,character-method}
\alias{localResultNames,SpatialFeatureExperiment-method}
\alias{localResultNames<-,SpatialFeatureExperiment,character-method}
\alias{localResult,SpatialFeatureExperiment,missing-method}
\alias{localResult,SpatialFeatureExperiment,numeric-method}
\alias{localResult,SpatialFeatureExperiment,character-method}
\alias{localResult<-,SpatialFeatureExperiment,missing-method}
\alias{localResult<-,SpatialFeatureExperiment,numeric-method}
\alias{localResult<-,SpatialFeatureExperiment,character-method}
\title{Organize results from local spatial statistics}
\usage{
\S4method{localResults}{SpatialFeatureExperiment,missing,missing}(
  x,
  sample_id = NULL,
  name,
  features = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  withDimnames = TRUE,
  ...
)

\S4method{localResults}{SpatialFeatureExperiment,missing,missing}(
  x,
  sample_id = NULL,
  name,
  features = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  withDimnames = TRUE,
  ...
) <- value

\S4method{localResults}{SpatialFeatureExperiment,ANY,character}(
  x,
  sample_id = NULL,
  name,
  features = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  withDimnames = TRUE,
  ...
)

\S4method{localResults}{SpatialFeatureExperiment,ANY,character}(
  x,
  sample_id = NULL,
  name,
  features = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  withDimnames = TRUE,
  ...
) <- value

\S4method{localResultNames}{SpatialFeatureExperiment}(x)

\S4method{localResultNames}{SpatialFeatureExperiment,character}(x) <- value

\S4method{localResult}{SpatialFeatureExperiment,missing}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
)

\S4method{localResult}{SpatialFeatureExperiment,numeric}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
)

\S4method{localResult}{SpatialFeatureExperiment,character}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
)

\S4method{localResult}{SpatialFeatureExperiment,missing}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
) <- value

\S4method{localResult}{SpatialFeatureExperiment,numeric}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
) <- value

\S4method{localResult}{SpatialFeatureExperiment,character}(
  x,
  type,
  feature,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  sample_id = NULL,
  withDimnames = TRUE
) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{sample_id}{Sample ID to get or set geometries.}

\item{name}{Name of the graphs to add to each sample_id; used in the
\code{spatialGraphs} replacement method as it must be character while
\code{type} can be either an integer index or a name.}

\item{features}{Features whose local results to get or set, for
\code{localResults} getter and setter for multiple features at a time.}

\item{colGeometryName}{Which \code{colGeometry} to get or set local results.}

\item{annotGeometryName}{Which \code{annotGeometry} to get or set local
results.}

\item{withDimnames}{Logical. If \code{TRUE}, then the dimnames (colnames or
rownames) of the gene count matrix should correspond to row names of the
\code{sf} data frames of interest.}

\item{...}{Ignored}

\item{value}{Values to set, should be either a matrix or a data frame.}

\item{type}{An integer specifying the index or string specifying the name of
the *Geometry to query or replace. If missing, then the first item in the
*Geometries will be returned or replaced.}

\item{feature}{Feature whose local results to get or set, for
\code{localResult} getter and setter for one feature at a time.}
}
\value{
\code{localResults} returns a named list each element of which is a
  set of local results of interest. \code{localResult} returns a matrix or a
  data frame, whichever the original is when it's set.
  \code{localResultNames} returns a character vector. Setters return an SFE
  object with the desired field set. For genes and \code{colData} columns,
  the local results are stored in the \code{localResults} field in
  \code{int_colData}, whereas for \code{colGeometries} and
  \code{annotGeometries}, the local results are stored as columns in the same
  \code{sf} data frames.
}
\description{
Local spatial statics like local Moran's I, local Geary's C, Getis-Ord Gi*,
and geographically weighted PCA loadings return values at each spatial
location. Just like dimension reductions, these results are clearly
associated with the braoder SFE object, so they should have a place within
the object. The \code{localResults} field in the SFE object stores these
results that has a value for each spatial location.
}
\examples{
# Toy example
sfe <- readRDS(system.file("testdata/sfe_toy.rds", package = "SpatialFeatureExperiment"))
set.seed(29)
toy_res1 <- matrix(rnorm(10), nrow = 5, ncol = 2,
                   dimnames = list(colnames(sfe), c("meow", "purr")))
toy_res2 <- matrix(rpois(10, lambda = 2), nrow = 5, ncol = 2,
                   dimnames = list(colnames(sfe), c("sassy", "tortitude")))
# Need to rewrite these examples
}
