% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{mirrorImg}
\alias{mirrorImg}
\alias{mirrorImg,BioFormatsImage-method}
\alias{mirrorImg,EBImage-method}
\title{Mirror/flip images}
\usage{
\S4method{mirrorImg}{SpatRasterImage}(x, direction = c("vertical", "horizontal"), ...)

\S4method{mirrorImg}{BioFormatsImage}(x, direction = c("vertical", "horizontal"), resolution = 4L)

\S4method{mirrorImg}{EBImage}(x, direction = c("vertical", "horizontal"), ...)
}
\arguments{
\item{x}{SpatRaster or SpatVector}

\item{direction}{character. Should (partially) match "vertical" to flip by rows, or "horizontal" to flip by columns}

\item{...}{additional arguments for writing files as in \code{\link[terra]{writeRaster}}}

\item{resolution}{Resolution to read in from OME-TIFF, defaults to 4, which
is a medium resolution in Xenium.}
}
\value{
For \code{SpatRasterImage} and \code{EBImage}, object of the same
  class. For \code{BioFormatsImage}, the image of the specified resolution is
  read into memory and then the \code{EBImage} method is called, returning
  \code{EBImage}. The extent is not changed.
}
\description{
Flip images along the middle horizontal or vertical axis.
}
\seealso{
Other image methods: 
\code{\link{SFE-image}},
\code{\link{cropImg}()},
\code{\link{imgRaster}()},
\code{\link{imgSource}()},
\code{\link{rotateImg}()},
\code{\link{transposeImg}()}
}
\concept{Image and raster}
\concept{image methods}
