% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read10xVisiumSFE}
\alias{read10xVisiumSFE}
\title{Read 10X Visium data as SpatialFeatureExperiment}
\usage{
read10xVisiumSFE(
  dirs = NULL,
  sample_id = paste0("sample", sprintf("\%02d", seq_along(dirs))),
  type = c("HDF5", "sparse"),
  data = c("filtered", "raw"),
  images = c("lowres", "hires"),
  unit = c("full_res_image_pixel", "micron"),
  style = "W",
  zero.policy = NULL,
  row.names = c("id", "symbol"),
  flip = c("geometry", "image", "none"),
  read_spatial_enrichment = TRUE
)
}
\arguments{
\item{dirs}{Directory for each sample that contains the \code{spatial} and
\code{raw/filtered_featues_bc_matrix} directories.}

\item{sample_id}{Which sample(s) in the SFE object to use for the graph. Can
also be "all", which means this function will compute the graph for all
samples independently.}

\item{type}{Either "HDF5", and the matrix will be represented as
\code{TENxMatrix}, or "sparse", and the matrix will be read as a
\code{dgCMatrix}.}

\item{data}{character string specifying whether to read in
filtered (spots mapped to tissue) or raw data (all spots).}

\item{images}{character vector specifying which images to include. 
Valid values are \code{"lowres", "hires", "fullres", "detected", "aligned"}}

\item{unit}{Whether to use pixels in full resolution image or microns as the
unit. If using microns, then spacing between spots in pixels will be used
to convert the coordinates into microns, as the spacing is known to be 100
microns. This is used to plot scale bar.}

\item{style}{\code{style} can take values \dQuote{W}, \dQuote{B}, \dQuote{C}, \dQuote{U}, \dQuote{minmax} and \dQuote{S}}

\item{zero.policy}{default NULL, use global option value; if FALSE stop with error for any empty neighbour sets, if TRUE permit the weights list to be formed with zero-length weights vectors}

\item{row.names}{String specifying whether to use Ensembl IDs ("ID") or gene symbols ("Symbol") as row names. 
For symbols, the Ensembl ID will be appended to disambiguate rows where the same symbol corresponds to multiple Ensembl IDs.}

\item{flip}{Whether to flip the geometries or the images, because in
\code{sf} and \code{terra}, the geometries use the Cartesian coordinates
greater y coordinates going up, while in images, greater y values go down.
Originally the Visium spots are in pixels in full res image. Either the
image or the geometry needs to be flipped for them match in the Cartesian
coordinate system.}

\item{read_spatial_enrichment}{Logical, whether to read the
`spatial_enrichment.csv` file from Visium output if the file is present and
add its contents to `rowData`.}
}
\value{
A SpatialFeatureExperiment object. The images might need to be
  manually transposed and/or mirrored to match the spots in this version of
  this package.
}
\description{
Read Space Ranger output from Visium v1 (not HD) as a
SpatialFeatureExperiment object, where spots are represented with polygons in
the colGeometry called "spotPoly". Other
geometries can be added later after the dataset is read. If \code{data =
"filtered"}, then spatial neighborhood graphs of the spots are also computed
and stored in the colGraph called "visium" in all samples for downstream
spatial analyses.
}
\note{
It is assumed that the images have not been cropped. Otherwise the
  images might not align with the spots.
}
\examples{
dir <- system.file("extdata", package = "SpatialFeatureExperiment")

sample_ids <- c("sample01", "sample02")
samples <- file.path(dir, sample_ids, "outs")

list.files(samples[1])
list.files(file.path(samples[1], "spatial"))
(sfe <- read10xVisiumSFE(dirs = samples, sample_id = sample_ids,
    type = "sparse", data = "filtered"
))

}
\concept{Read data into SFE}
