% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read10xVisiumSFE}
\alias{read10xVisiumSFE}
\title{Read 10X Visium data as SpatialFeatureExperiment}
\usage{
read10xVisiumSFE(
  samples = "",
  sample_id = paste0("sample", sprintf("\%02d", seq_along(samples))),
  type = c("HDF5", "sparse"),
  data = c("filtered", "raw"),
  images = "lowres",
  load = TRUE,
  style = "W",
  zero.policy = NULL
)
}
\arguments{
\item{samples}{a character vector specifying one or more directories, 
each corresponding to a 10x Genomics Visium sample (see details);
if provided, names will be used as sample identifiers}

\item{sample_id}{character string specifying unique sample identifiers,
one for each directory specified via \code{samples}; 
ignored if \code{!is.null(names(samples))}}

\item{type}{character string specifying 
the type of format to read count data from
(see \code{\link{read10xCounts}})}

\item{data}{character string specifying whether to read in
filtered (spots mapped to tissue) or raw data (all spots).}

\item{images}{character vector specifying which images to include. 
Valid values are \code{"lowres", "hires", "fullres", "detected", "aligned"}}

\item{load}{logical; should the image(s) be loaded into memory
as a \code{grob}? If FALSE, will store the path/URL instead.}

\item{style}{\code{style} can take values \dQuote{W}, \dQuote{B}, \dQuote{C}, \dQuote{U}, \dQuote{minmax} and \dQuote{S}}

\item{zero.policy}{default NULL, use global option value; if FALSE stop with error for any empty neighbour sets, if TRUE permit the weights list to be formed with zero-length weights vectors}
}
\value{
A SpatialFeatureExperiment object
}
\description{
Read Space Ranger output as a SpatialFeatureExperiment object, where spots
are represented with polygons in the colGeometry called "spotPoly". Other
geometries can be added later after the dataset is read. If \code{data =
"filtered"}, then spatial neighborhood graphs of the spots are also computed
and stored in the colGraph called "visium" in all samples for downstream
spatial analyses.
}
\examples{
library(SpatialExperiment) # Just for the example directory
dir <- system.file(
  file.path("extdata", "10xVisium"),
  package = "SpatialExperiment")

sample_ids <- c("section1", "section2")
samples <- file.path(dir, sample_ids, "outs")

list.files(samples[1])
list.files(file.path(samples[1], "spatial"))
file.path(samples[1], "raw_feature_bc_matrix")
(sfe <- read10xVisiumSFE(samples, sample_ids, type = "sparse", data = "raw",
                         load = FALSE))
}
