% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readVizgen}
\alias{readVizgen}
\title{Read Vizgen MERFISH output as SpatialFeatureExperiment}
\usage{
readVizgen(
  data_dir,
  z = 3L,
  use_cellpose = TRUE,
  sample_id = "sample01",
  min_area = 15,
  image = c("DAPI", "PolyT"),
  flip = c("geometry", "image", "none"),
  max_flip = "50 MB",
  BPPARAM = SerialParam()
)
}
\arguments{
\item{data_dir}{Top level directory of Vizgen output, which contains
directories \code{cell_boundaries} and \code{images}, and files
\code{cell_by_gene.csv}, \code{cell_metadata.csv}, and
\code{detected_transcripts.csv}.}

\item{z}{Index of z plane to read.}

\item{use_cellpose}{Logical, whether to use Cellpose parquet files if
present.}

\item{sample_id}{A \code{character} sample identifier, which matches the
\code{sample_id} in \code{\link[SpatialExperiment]{imgData}}. The \code{sample_id} will also
be stored in a new column in \code{\link[SpatialExperiment]{colData}}, if not already present.
Default = \code{sample01}.}

\item{min_area}{Minimum cell area in square microns. Anything smaller will be
considered artifact or debris and removed.}

\item{image}{Which image(s) to load, can be "DAPI", "PolyT", or both.}

\item{flip}{To flip the image, geometry coordinates, or none. Because the
image has the origin at the top left while the geometry has origin at the
bottom left, one of them needs to be flipped for them to match. If one of
them is already flipped, then use "none". The image will not be flipped if
it's GeoTIFF.}

\item{max_flip}{Maximum size of the image allowed to flip the image. Because
the image will be loaded into memory to be flipped. If the image is larger
than this size then the coordinates will be flipped instead.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying parallel
processing backend and number of threads to use to load cell segmentation
from HDF5 files from different fields of view (FOVs) with multiple cores. A
progress bar can be configured in the \code{\link{BiocParallelParam}}
object. When there are numerous FOVs, reading in the geometries can be time
consuming, so we recommend using a server and larger number of threads.
This argument is not used if \code{use_cellpose = TRUE} and the parquet
file is present.}
}
\value{
A \code{SpatialFeatureExperiment} object.
}
\description{
This function reads the standard Vizgen MERFISH output into an SFE object.
The coordinates are in microns. Cell centroids are read into
\code{\link{colGeometry}} "centroids", and cell segmentations are read into
\code{colGeometry} "cellSeg". The image(s) (polyT and DAPI) are also read as
\code{\link{SpatRaster}} objects so they are not loaded into memory unless
necessary. Because the image's origin is the top left while the geometry's
origin is bottom left, either the image or the geometry needs to be flipped.
Because the image accompanying MERFISH datasets are usually very large, the
coordinates will be flipped so the flipping operation won't load the entire
image into memory.
}
\examples{
dir_use <- system.file("extdata/vizgen", package = "SpatialFeatureExperiment")
sfe <- readVizgen(dir_use, z = 0L, use_cellpose = TRUE, image = "PolyT",
flip = "geometry")
}
