% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{colData}
\alias{rowData}
\alias{colData<-}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{spatialCoordsNames}
\alias{getImg}
\alias{imgData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SpatialExperiment}{\code{\link[SpatialExperiment:imgData-methods]{getImg}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{imgData}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoords}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoords<-}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoordsNames}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}}
}}

