% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{removeEmptySpace}
\alias{removeEmptySpace}
\title{Remove empty space}
\usage{
removeEmptySpace(sfe, sample_id = "all")
}
\arguments{
\item{sample_id}{Samples to crop. Can be multiple samples, or "all", which
means all samples. For multiple samples, \code{y} may have column
\code{sample_id} indicating which geometry subsets which sample. Only
samples included in the \code{sample_id} column are subsetted. If there is
no \code{sample_id} column or \code{y} is not specified, then the same
geometry or bounding box is used to subset all samples specified in the
\code{sample_id} argument.}
}
\value{
An SFE object with empty space removed.
}
\description{
For each sample independently, all geometries and \code{spatialCoords} are
translated so the origin is at the minimum coordinates of the bounding box
of all geometries of the sample. This way coordinates of different samples
will be more comparable.
}
\note{
Unlike other functions in this package, this function operates on all
samples by default.
}
