% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{rotateImg}
\alias{rotateImg}
\alias{rotateImg,SpatRasterImage-method}
\alias{rotateImg,BioFormatsImage-method}
\alias{rotateImg,EBImage-method}
\title{Rotate image}
\usage{
\S4method{rotateImg}{SpatRasterImage}(x, degrees, maxcell = 1e+07, ...)

\S4method{rotateImg}{BioFormatsImage}(x, degrees, ...)

\S4method{rotateImg}{EBImage}(x, degrees, ...)
}
\arguments{
\item{x}{An object of class \code{*Image} as implemented in this package.}

\item{degrees}{How many degrees to rotate. Positive
number means clockwise and negative number means counterclockwise.}

\item{maxcell}{Max number of pixels to load \code{SpatRasterImage} into
memory. The default 1e7 is chosen because this is the approximate number of
pixels in the medium resolution image at \code{resolution = 4L} in Xenium
OME-TIFF to make different methods of this function consistent.}

\item{...}{Ignored. It's there so different methods can all be passed to the
same \code{lapply} in the method for SFE objects. Some methods have extra
arguments.}
}
\value{
An \code{EBImage} object. Both \code{SpatRasterImage} and
  \code{BioFormatsImage} will be loaded into memory as \code{EBImage} before
  rotating.
}
\description{
As in \code{SpatialExperiment}, rotation here must be a multiple of 90
degrees.
}
\seealso{
Other image methods: 
\code{\link{SFE-image}},
\code{\link{cropImg}()},
\code{\link{imgRaster}()},
\code{\link{imgSource}()},
\code{\link{mirrorImg}()},
\code{\link{transposeImg}()}
}
\concept{Image and raster}
\concept{image methods}
