% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateMinRect.R
\name{rotateMinRect}
\alias{rotateMinRect}
\alias{rotateMinRect,matrix-method}
\alias{rotateMinRect,sfg-method}
\alias{rotateMinRect,sfc-method}
\alias{rotateMinRect,sf-method}
\title{Rotate the tissue to reduce empty space in plots}
\usage{
\S4method{rotateMinRect}{matrix}(x, orientation = c("horizontal", "vertical"))

\S4method{rotateMinRect}{sfg}(x, orientation = c("horizontal", "vertical"))

\S4method{rotateMinRect}{sfc}(x, orientation = c("horizontal", "vertical"))

\S4method{rotateMinRect}{sf}(x, orientation = c("horizontal", "vertical"))
}
\arguments{
\item{x}{Object of interest to be rotated, can be a matrix with 2 columns,
\code{sf}, \code{sfc}, \code{sfg}, or \code{SpatialFeatureExperiment}.}

\item{orientation}{Whether you want the rotated sample to be horizontal or
vertical.}
}
\value{
Object of the same class as \code{x}, rotated.
}
\description{
This function finds the minimum bounding rotated rectangle and then rotates
the sample so the minimum bounding rotated rectangle is horizontal or
vertical. This will reduce empty space in plots when the sample has a long
shape.
}
