% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{scaleImg}
\alias{scaleImg}
\alias{scaleImg,SpatRasterImage-method}
\alias{scaleImg,BioFormatsImage-method}
\alias{scaleImg,EBImage-method}
\title{Scale image}
\usage{
\S4method{scaleImg}{SpatRasterImage}(x, factor, ...)

\S4method{scaleImg}{BioFormatsImage}(x, factor, ...)

\S4method{scaleImg}{EBImage}(x, factor, ...)
}
\arguments{
\item{x}{An object of class \code{*Image} as implemented in this package.}

\item{factor}{Numeric, scaling factor.}

\item{...}{Ignored. It's there so different methods can all be passed to the
same \code{lapply} in the method for SFE objects. Some methods have extra
arguments.}
}
\value{
A \code{*Image} object of the same class that has been scaled. Behind
the scene, it's only the extent that has been changed and the images are not
changed. The center of the image is unchanged.
}
\description{
This function scales the image about its center. After scaling, the center
of the image is not shifted.
}
