% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{SFE-image}
\alias{SFE-image}
\alias{addImg,SpatialFeatureExperiment-method}
\alias{transposeImg,SpatialFeatureExperiment-method}
\alias{mirrorImg,SpatialFeatureExperiment-method}
\alias{rotateImg,SpatialFeatureExperiment-method}
\alias{translateImg,SpatialFeatureExperiment-method}
\title{Methods for handling image-related data}
\usage{
\S4method{addImg}{SpatialFeatureExperiment}(
  x,
  imageSource,
  sample_id,
  image_id,
  extent = NULL,
  scale_fct = 1,
  file = deprecated()
)

\S4method{transposeImg}{SpatialFeatureExperiment}(x, sample_id = NULL, image_id = NULL, resolution = 4L)

\S4method{mirrorImg}{SpatialFeatureExperiment}(
  x,
  sample_id = NULL,
  image_id = NULL,
  direction = "vertical",
  resolution = 4L
)

\S4method{rotateImg}{SpatialFeatureExperiment}(
  x,
  sample_id = NULL,
  image_id = NULL,
  degrees,
  resolution = 4L,
  maxcell = 1e+07
)

\S4method{translateImg}{SpatialFeatureExperiment}(x, sample_id = NULL, image_id = NULL, v)
}
\arguments{
\item{x}{A SFE object.}

\item{sample_id}{Which sample the image is associated with. Use
\code{\link{sampleIDs}} to get sample IDs present in the SFE object.}

\item{image_id}{Image ID, such as "lowres" and "hires" for Visium data and
"DAPI" and "PolyT" for Vizgen MERFISH data.}

\item{extent}{A numeric vector of length 4 with names of the set xmin, ymin,
xmax, and ymax, specifying the extent of the image.}

\item{scale_fct}{Scale factor -- multiply pixel coordinates in full
resolution image by this scale factor should yield pixel coordinates in a
different resolution. \code{extent} takes precedence over \code{scale_fct}.}

\item{file}{File from which to read the image.}

\item{resolution}{Resolution to read in from OME-TIFF, defaults to 4, which
is a medium resolution in Xenium.}

\item{direction}{character. Should (partially) match "vertical" to flip by rows, or "horizontal" to flip by columns}

\item{degrees}{How many degrees to rotate, must be multiples of 90. Positive
number means clockwise and negative number means counterclockwise.}

\item{maxcell}{Max number of pixels to load \code{SpatRasterImage} into
memory. The default 1e7 is chosen because this is the approximate number of
pixels in the medium resolution image at \code{resolution = 4L} in Xenium
OME-TIFF to make different methods of this function consistent.}
}
\description{
Generics of these functions are defined in \code{SpatialExperiment}, except
for \code{transposeImg}. These SFE methods cater to the new image-related
classes in SFE. The SPE method for \code{getImg}, \code{rmvImg}, and
\code{imgRaster} don't need to be modified for SFE and are hence not
implemented here, but are simply re-exported.
}
\details{
Method of \code{\link{transposeImg}}, \code{\link{mirrorImg}}, and
\code{\link{rotateImg}} perform the method on all images within the SFE
object that are specified with \code{sample_id} and \code{image_id}.
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData("small")
img_path <- system.file(file.path("extdata", "sample01", "outs", "spatial",
"tissue_lowres_image.png"), package = "SpatialFeatureExperiment")
sfe <- addImg(sfe, img_path, sample_id = "Vis5A", image_id = "lowres", scale_fct =
0.023)
img <- getImg(sfe)
# SpatRasterImage method
img_t <- transposeImg(img)
# SFE method
sfe <- transposeImg(sfe, sample_id = "Vis5A", image_id = "lowres")
}
\seealso{
Other image methods: 
\code{\link{cropImg}()},
\code{\link{imgRaster}()},
\code{\link{imgSource}()},
\code{\link{mirrorImg}()},
\code{\link{rotateImg}()},
\code{\link{transposeImg}()}
}
\concept{Image and raster}
\concept{image methods}
