% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{SFE-image}
\alias{SFE-image}
\alias{transposeImg}
\alias{SpatRasterImage-class}
\alias{addImg,SpatialFeatureExperiment-method}
\alias{transposeImg,SpatRasterImage-method}
\alias{mirrorImg,SpatRasterImage-method}
\alias{transposeImg,SpatialFeatureExperiment-method}
\alias{mirrorImg,SpatialFeatureExperiment-method}
\alias{imgRaster,SpatRasterImage-method}
\alias{imgSource,SpatRasterImage-method}
\title{Images in SpatialFeatureExperiment object}
\usage{
\S4method{addImg}{SpatialFeatureExperiment}(x, file, sample_id, image_id, extent = NULL, scale_fct = 1)

\S4method{transposeImg}{SpatRasterImage}(x)

\S4method{mirrorImg}{SpatRasterImage}(x, direction = "vertical")

\S4method{transposeImg}{SpatialFeatureExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{mirrorImg}{SpatialFeatureExperiment}(x, sample_id = NULL, image_id = NULL, direction = "vertical")

\S4method{imgRaster}{SpatRasterImage}(x)

\S4method{imgSource}{SpatRasterImage}(x)
}
\arguments{
\item{x}{SpatRaster or SpatVector}

\item{file}{File from which to read the image.}

\item{sample_id}{Which sample the image is associated with. Use
\code{\link{sampleIDs}} to get sample IDs present in the SFE object.}

\item{image_id}{Image ID, such as "lowres" and "hires" for Visium data and
"DAPI" and "PolyT" for Vizgen MERFISH data.}

\item{extent}{A numeric vector of length 4 with names of the set xmin, ymin,
xmax, and ymax, specifying the extent of the image.}

\item{scale_fct}{Scale factor -- multiply pixel coordinates in full
resolution image by this scale factor should yield pixel coordinates in a
different resolution. \code{extent} takes precedence over \code{scale_fct}.}

\item{direction}{character. Should (partially) match "vertical" to flip by rows, or "horizontal" to flip by columns}
}
\value{
Methods for \code{SpatRasterImage} return a modified
  \code{SpatRasterImage}, and methods for SFE return a modified SFE object.
}
\description{
\code{SpatialFeatureExperiment} and the \code{Voyager} package work with
images differently from \code{SpatialExperiment}. In SFE and
\code{Voyager}'s, plotting functions for SFE objects, the images are read
with \code{\link{rast}} and represented as \code{SpatRaster}, so the image is
not entirely loaded into memory unless necessary. Plotting will not load a
large image into memory; rather the image will be downsampled and the
downsampled version is plotted.
}
\note{
If the image is already a GeoTIFF file that already has an extent, then
  the extent associated with the file will be honored and the \code{extent}
  and \code{scale_fct} arguments are ignored. Also, when the image is
  transposed, it is flipped about the axis going from top left to bottom
  right.
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData("small")
img_path <- system.file(file.path("extdata", "sample01", "outs", "spatial",
"tissue_lowres_image.png"),
package = "SpatialFeatureExperiment")
sfe <- addImg(sfe, img_path, sample_id = "Vis5A", image_id = "lowres",
scale_fct = 0.023)
img <- getImg(sfe)
# SpatRasterImage method
img_t <- transposeImg(img)
# SFE method
sfe <- transposeImg(sfe, sample_id = "Vis5A", image_id = "lowres")
}
