% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE-class.R
\docType{class}
\name{SpatialFeatureExperiment-class}
\alias{SpatialFeatureExperiment-class}
\title{The SpatialFeatureExperiment class}
\description{
This class inherits from the \code{\link[SpatialExperiment]{SpatialExperiment}} (SPE) class,
which in turn inherits from \code{\link[SingleCellExperiment]{SingleCellExperiment}} (SCE).
\code{SpatialFeatureExperiment} stores geometries of spots or cells in
\code{sf} objects which form columns of a \code{DataFrame} which is in turn a
column of the \code{int_colData} \code{DataFrame} of the underlying SCE
object, just like \code{reducedDim} in SCE. Geometries of the tissue outline,
pathologist annotations, and objects (e.g. nuclei segmentation in a Visium
dataset) are stored in \code{sf} objects in a named list called
\code{annotGeometries} in \code{int_metadata}.
}
\concept{SpatialFeatureExperiment class}
