% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{SpatialFeatureExperiment-coercion}
\alias{SpatialFeatureExperiment-coercion}
\alias{toSpatialFeatureExperiment,SpatialExperiment-method}
\title{SpatialFeatureExperiment coercion methods}
\usage{
\S4method{toSpatialFeatureExperiment}{SpatialExperiment}(
  x,
  colGeometries = NULL,
  rowGeometries = NULL,
  annotGeometries = NULL,
  unit = NULL
)
}
\arguments{
\item{unit}{Unit the coordinates are in. I'm thinking about using some custom
engineering CRS's which can convert units and invert the y axis for
Cartesian vs. image orientations. Units are also helpful when plotting
scale bars. Ignored for now, until I find a better way to deal with it.}
}
\description{
The \code{SpatialFeatureExperiment} class inherits from
\code{SpatialExperiment}, which in turn inherits from
\code{SingleCellExperiment}. A \code{SpatialExperiment} object with
geometries in \code{colGeometries} in the \code{int_colData},
\code{rowGeometries} in the \code{int_elementMetadata}, or
\code{annotGeometries} in the \code{int_metadata} can be directly converted
to \code{SpatialFeatureExperiment} with \code{as(spe,
"SpatialFeatureExperiment")}. A \code{SpatialExperiment} object without the
geometries can also be converted; the coordinates in the \code{spatialCoords}
field will be used to make POINT geometries named "centroids" to add to
\code{colGeometries}. The geometries can also be supplied separately when
using \code{toSpatialFeatureExperiment}. For now coercion only works for
\code{SpatialExperiment}. I'll deal with \code{Seurat} and
\code{SingleCellExperiment} later.
}
