% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SpatialFeatureExperiment-subset}
\alias{SpatialFeatureExperiment-subset}
\alias{[,SpatialFeatureExperiment,ANY,ANY,ANY-method}
\title{Subsetting SpatialFeatureExperiment objects}
\usage{
\S4method{[}{SpatialFeatureExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{i}{Row indices for subsetting.}

\item{j}{column indices for subsetting.}

\item{...}{Passed to the \code{SingleCellExperiment} method of \code{[}.}

\item{drop}{Ignored as of version 1.9.2.}
}
\value{
A subsetted \code{SpatialFeatureExperiment} object.
}
\description{
Note that spatial neighborhood graphs may change meaning after subsetting.
For example, for a k nearest neighbor graph, after subsetting, some cells
might no longer have all k nearest neighbors from the original. The edge
weights will be recomputed from the binary neighborhood indicator with the
same normalization style as the original graph, such as "W" for row
normalization. When distance-based edge weights are used instead of the
binary indicator, the edge weights will be re-normalized, which is mostly
some rescaling. This should give the same results as recomputing the distance
based edge weights for styles "raw", "W", and "B" since the distances
themselves don't change, but the effects of other more complicated styles of
re-normalization on spatial statistics should be further investigated.
}
\examples{
# Just like subsetting matrices and SingleCellExperiment
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
sfe_subset <- sfe[seq_len(10), seq_len(10), drop = TRUE]
# Gives warning as graph reconstruction fails
\donttest{
sfe_subset <- sfe[seq_len(10), seq_len(10)]
}
}
\concept{Non-spatial operations}
