% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SpatialFeatureExperiment-subset}
\alias{SpatialFeatureExperiment-subset}
\title{Subsetting SpatialFeatureExperiment objects}
\usage{
\S4method{[}{SpatialFeatureExperiment,ANY,ANY,ANY}(x, i, j, ..., reconstruct_graphs = TRUE, drop = FALSE)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{i}{Row indices for subsetting.}

\item{j}{column indices for subsetting.}

\item{reconstruct_graph}{Logical, whether to reconstruct graph. If \code{FALSE},
then the old graphs will be kept and a warning will be issued that the node
indices in the graphs are no longer valid. At present, this only works with
the wrapper functions in this package that take in SFE objects and records
the info required to reconstruc the graphs.}
}
\value{
A subsetted \code{SpatialFeatureExperiment} object.
}
\description{
The method for SFE reconstructs the spatial graphs when the SFE object is
subsetted as the \code{listw} objects encodes the nodes with indices which
are no longer valid after subsetting as some nodes are no longer present.
}
