% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SpatialFeatureExperiment-subset}
\alias{SpatialFeatureExperiment-subset}
\alias{[,SpatialFeatureExperiment,ANY,ANY,ANY-method}
\title{Subsetting SpatialFeatureExperiment objects}
\usage{
\S4method{[}{SpatialFeatureExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{i}{Row indices for subsetting.}

\item{j}{column indices for subsetting.}

\item{...}{Passed to the \code{SingleCellExperiment} method of \code{[}.}

\item{drop}{Logical. If \code{FALSE}, then a warning will be issued that the
node indices in the graphs are no longer valid so the row and col graphs
affected by subsetting are dropped. At present, this only works with the
wrapper functions in this package that take in SFE objects and records the
info required to reconstruct the graphs. While this argument is ignored for
\code{SummarizedExperiment}}
}
\value{
A subsetted \code{SpatialFeatureExperiment} object.
}
\description{
The method for SFE reconstructs the spatial graphs when the SFE object is
subsetted as the \code{listw} objects encodes the nodes with indices which
are no longer valid after subsetting as some nodes are no longer present.
}
\examples{
# Just like subsetting matrices and SingleCellExperiment
library(SFEData)
sfe <- McKellarMuscleData(dataset = "small")
sfe_subset <- sfe[1:10, 1:10, drop = TRUE]
# Gives warning as graph reconstruction fails
\donttest{
sfe_subset <- sfe[1:10, 1:10]
}
}
