% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\docType{methods}
\name{spatialGraphs}
\alias{spatialGraphs}
\alias{spatialGraphs,SpatialFeatureExperiment-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment-method}
\alias{spatialGraph,SpatialFeatureExperiment,missing-method}
\alias{spatialGraph,SpatialFeatureExperiment,numeric-method}
\alias{spatialGraph,SpatialFeatureExperiment,character-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,missing,numeric,listw-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,numeric,numeric,listw-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,numeric,listw-method}
\title{Spatial graph methods}
\usage{
\S4method{spatialGraphs}{SpatialFeatureExperiment}(x, MARGIN)

\S4method{spatialGraphs}{SpatialFeatureExperiment}(x, MARGIN) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,missing}(x, type, MARGIN)

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric}(x, type, MARGIN)

\S4method{spatialGraph}{SpatialFeatureExperiment,character}(x, type, MARGIN)

\S4method{spatialGraph}{SpatialFeatureExperiment,missing,numeric,listw}(x, type, MARGIN) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric,numeric,listw}(x, type, MARGIN) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,numeric,listw}(x, type, MARGIN) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{MARGIN}{As in \code{\link{apply}}. 1 stands for rows and 2 stands for
columns. In addition, 3 stands for spatial neighborhood graphs that
correspond to \code{annotGeometries}.}

\item{value}{A \code{listw} object (\code{*Graph}), or a named list of
\code{listw} objects (\code{*Graphs}).}

\item{type}{An integer specifying the index or string specifying the name of
the *Graph to query or replace. If missing, then the first item in the
*Graph will be returned or replaced.}
}
\description{
Spatial neighborhood graphs as \code{spdep}'s \code{listw} objects are stored
in the \code{int_metadata} of the SFE object. The \code{listw} class is used
because \code{spdep} has many useful methods that rely on the neighborhood
graph as \code{listw}.
}
