% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\name{spatialGraphs}
\alias{spatialGraphs}
\alias{spatialGraphs,SpatialFeatureExperiment-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment-method}
\alias{spatialGraph,SpatialFeatureExperiment,missing-method}
\alias{spatialGraph,SpatialFeatureExperiment,numeric-method}
\alias{spatialGraph,SpatialFeatureExperiment,character-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,missing,listw-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,numeric,listw-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,listw-method}
\title{Spatial graph methods}
\usage{
\S4method{spatialGraphs}{SpatialFeatureExperiment}(x)

\S4method{spatialGraphs}{SpatialFeatureExperiment}(x) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,missing}(x, type)

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric}(x, type)

\S4method{spatialGraph}{SpatialFeatureExperiment,character}(x, type)

\S4method{spatialGraph}{SpatialFeatureExperiment,missing,listw}(x, type) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric,listw}(x, type) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,listw}(x, type) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}
}
\description{
Spatial neighborhood graphs as \code{spdep}'s \code{listw} objects are stored
in the \code{int_metadata} of the SFE object. The \code{listw} class is used
because \code{spdep} has many useful methods that rely on the neighborhood
graph as \code{listw}.
}
