% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\docType{methods}
\name{spatialGraphs}
\alias{spatialGraphs}
\alias{spatialGraphs,SpatialFeatureExperiment,missing,numeric-method}
\alias{colGraphs,SpatialFeatureExperiment,missing-method}
\alias{rowGraphs,SpatialFeatureExperiment,missing-method}
\alias{annotGraphs,SpatialFeatureExperiment,missing-method}
\alias{spatialGraphs,SpatialFeatureExperiment,character,numeric-method}
\alias{colGraphs,SpatialFeatureExperiment,character-method}
\alias{rowGraphs,SpatialFeatureExperiment,character-method}
\alias{annotGraphs,SpatialFeatureExperiment,character-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,missing,numeric-method}
\alias{colGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{rowGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{annotGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,character,numeric-method}
\alias{colGraphs<-,SpatialFeatureExperiment,character-method}
\alias{rowGraphs<-,SpatialFeatureExperiment,character-method}
\alias{annotGraphs<-,SpatialFeatureExperiment,character-method}
\alias{spatialGraphNames,SpatialFeatureExperiment,character,numeric-method}
\alias{spatialGraphNames<-,SpatialFeatureExperiment,character,numeric,character-method}
\alias{colGraphNames}
\alias{rowGraphNames}
\alias{annotGraphNames}
\alias{colGraphNames<-}
\alias{rowGraphNames<-}
\alias{annotGraphNames<-}
\alias{spatialGraph,SpatialFeatureExperiment,character,missing-method}
\alias{spatialGraph,SpatialFeatureExperiment,character,numeric-method}
\alias{spatialGraph,SpatialFeatureExperiment,character,character-method}
\alias{colGraph}
\alias{rowGraph}
\alias{annotGraph}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,missing,numeric,listw-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,numeric,numeric,ANY-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,character,numeric,ANY-method}
\alias{colGraph<-}
\alias{rowGraph<-}
\alias{annotGraph<-}
\title{Spatial graph methods}
\usage{
\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,numeric}(x, sample_id, MARGIN)

\S4method{colGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{rowGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{annotGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{spatialGraphs}{SpatialFeatureExperiment,character,numeric}(x, sample_id, MARGIN)

\S4method{colGraphs}{SpatialFeatureExperiment,character}(x, sample_id)

\S4method{rowGraphs}{SpatialFeatureExperiment,character}(x, sample_id)

\S4method{annotGraphs}{SpatialFeatureExperiment,character}(x, sample_id)

\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,numeric}(x, sample_id, MARGIN) <- value

\S4method{colGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{rowGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{annotGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{spatialGraphs}{SpatialFeatureExperiment,character,numeric}(x, sample_id, MARGIN) <- value

\S4method{colGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{rowGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{annotGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{spatialGraphNames}{SpatialFeatureExperiment,character,numeric}(x, sample_id, MARGIN)

\S4method{spatialGraphNames}{SpatialFeatureExperiment,character,numeric,character}(x, sample_id, MARGIN) <- value

colGraphNames(x, sample_id)

rowGraphNames(x, sample_id)

annotGraphNames(x, sample_id)

colGraphNames(x, sample_id) <- value

rowGraphNames(x, sample_id) <- value

annotGraphNames(x, sample_id) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,missing}(x, sample_id, type, MARGIN)

\S4method{spatialGraph}{SpatialFeatureExperiment,character,numeric}(x, sample_id, type, MARGIN)

\S4method{spatialGraph}{SpatialFeatureExperiment,character,character}(x, sample_id, type, MARGIN)

colGraph(x, sample_id, type = 1L)

rowGraph(x, sample_id, type = 1L)

annotGraph(x, sample_id, type = 1L)

\S4method{spatialGraph}{SpatialFeatureExperiment,character,missing,numeric,listw}(x, sample_id, type, MARGIN) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,numeric,numeric,ANY}(x, sample_id, type, MARGIN) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,character,numeric,ANY}(x, sample_id, type, MARGIN) <- value

colGraph(x, sample_id, type = 1L) <- value

rowGraph(x, sample_id, type = 1L) <- value

annotGraph(x, sample_id, type = 1L) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{sample_id}{Name of the sample the graph is associated with. This is
useful when multiple pieces of tissues are in the same SFE object (say for
a joint dimension reduction and clustering) and the spatial neighborhood is
only meaningful within the same piece of tissue. See the \code{sample_id}
argument in \code{\link{SpatialExperiment}}.}

\item{MARGIN}{As in \code{\link{apply}}. 1 stands for rows and 2 stands for
columns. In addition, 3 stands for spatial neighborhood graphs that
correspond to \code{annotGeometries}.}

\item{value}{A \code{listw} object (\code{*Graph}), or a named list of list
of \code{listw} objects (\code{*Graphs}) where the names of the top level
list are \code{sample_id}s when adding graphs for all samples in the margin
of interest, or a list of \code{listw} objects when adding graphs for one
sample in one margin.}

\item{type}{An integer specifying the index or string specifying the name of
the *Graph to query or replace. If missing, then the first item in the
*Graph will be returned or replaced.}
}
\description{
Spatial neighborhood graphs as \code{spdep}'s \code{listw} objects are stored
in the \code{int_metadata} of the SFE object. The \code{listw} class is used
because \code{spdep} has many useful methods that rely on the neighborhood
graph as \code{listw}.
}
