% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\docType{methods}
\name{spatialGraphs}
\alias{spatialGraphs}
\alias{rowGraphs}
\alias{rowGraphs<-}
\alias{spatialGraph}
\alias{spatialGraph<-}
\alias{spatialGraphNames}
\alias{colGraphs}
\alias{colGraphs<-}
\alias{spatialGraphNames<-}
\alias{spatialGraphs<-}
\alias{annotGraphs}
\alias{annotGraphs<-}
\alias{spatialGraphs,SpatialFeatureExperiment,missing,missing-method}
\alias{spatialGraphs,SpatialFeatureExperiment,numeric,missing-method}
\alias{spatialGraphs,SpatialFeatureExperiment,missing,character-method}
\alias{colGraphs,SpatialFeatureExperiment,missing-method}
\alias{rowGraphs,SpatialFeatureExperiment,missing-method}
\alias{annotGraphs,SpatialFeatureExperiment,missing-method}
\alias{spatialGraphs,SpatialFeatureExperiment,numeric,character-method}
\alias{colGraphs,SpatialFeatureExperiment,ANY-method}
\alias{rowGraphs,SpatialFeatureExperiment,ANY-method}
\alias{annotGraphs,SpatialFeatureExperiment,ANY-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,missing,missing-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,numeric,missing-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,numeric,NULL-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,missing,character-method}
\alias{colGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{rowGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{annotGraphs<-,SpatialFeatureExperiment,missing-method}
\alias{spatialGraphs<-,SpatialFeatureExperiment,numeric,character-method}
\alias{colGraphs<-,SpatialFeatureExperiment,character-method}
\alias{rowGraphs<-,SpatialFeatureExperiment,character-method}
\alias{annotGraphs<-,SpatialFeatureExperiment,character-method}
\alias{spatialGraphNames,SpatialFeatureExperiment,numeric-method}
\alias{spatialGraphNames<-,SpatialFeatureExperiment,numeric,ANY,character-method}
\alias{colGraphNames}
\alias{rowGraphNames}
\alias{annotGraphNames}
\alias{colGraphNames<-}
\alias{rowGraphNames<-}
\alias{annotGraphNames<-}
\alias{spatialGraph,SpatialFeatureExperiment,missing,numeric-method}
\alias{spatialGraph,SpatialFeatureExperiment,numeric,numeric-method}
\alias{spatialGraph,SpatialFeatureExperiment,character,numeric-method}
\alias{colGraph}
\alias{rowGraph}
\alias{annotGraph}
\alias{spatialGraph<-,SpatialFeatureExperiment,missing,numeric-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,numeric,numeric-method}
\alias{spatialGraph<-,SpatialFeatureExperiment,character,numeric-method}
\alias{colGraph<-}
\alias{rowGraph<-}
\alias{annotGraph<-}
\title{Spatial graph methods}
\usage{
\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,missing}(x, MARGIN, sample_id)

\S4method{spatialGraphs}{SpatialFeatureExperiment,numeric,missing}(x, MARGIN, sample_id)

\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,character}(x, MARGIN, sample_id = NULL)

\S4method{colGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{rowGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{annotGraphs}{SpatialFeatureExperiment,missing}(x, sample_id)

\S4method{spatialGraphs}{SpatialFeatureExperiment,numeric,character}(x, MARGIN, sample_id)

\S4method{colGraphs}{SpatialFeatureExperiment,ANY}(x, sample_id = NULL)

\S4method{rowGraphs}{SpatialFeatureExperiment,ANY}(x, sample_id = NULL)

\S4method{annotGraphs}{SpatialFeatureExperiment,ANY}(x, sample_id = NULL)

\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,missing}(x, MARGIN, sample_id) <- value

\S4method{spatialGraphs}{SpatialFeatureExperiment,numeric,missing}(x, MARGIN, sample_id) <- value

\S4method{spatialGraphs}{SpatialFeatureExperiment,numeric,`NULL`}(x, MARGIN, sample_id) <- value

\S4method{spatialGraphs}{SpatialFeatureExperiment,missing,character}(x, MARGIN, sample_id) <- value

\S4method{colGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{rowGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{annotGraphs}{SpatialFeatureExperiment,missing}(x, sample_id) <- value

\S4method{spatialGraphs}{SpatialFeatureExperiment,numeric,character}(x, MARGIN, sample_id) <- value

\S4method{colGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{rowGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{annotGraphs}{SpatialFeatureExperiment,character}(x, sample_id) <- value

\S4method{spatialGraphNames}{SpatialFeatureExperiment,numeric}(x, MARGIN, sample_id)

\S4method{spatialGraphNames}{SpatialFeatureExperiment,numeric,ANY,character}(x, MARGIN, sample_id = NULL) <- value

colGraphNames(x, sample_id = NULL)

rowGraphNames(x, sample_id = NULL)

annotGraphNames(x, sample_id = NULL)

colGraphNames(x, sample_id = NULL) <- value

rowGraphNames(x, sample_id = NULL) <- value

annotGraphNames(x, sample_id = NULL) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,missing,numeric}(x, type, MARGIN, sample_id = NULL)

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric,numeric}(x, type, MARGIN, sample_id = NULL)

\S4method{spatialGraph}{SpatialFeatureExperiment,character,numeric}(x, type, MARGIN, sample_id = NULL)

colGraph(x, type = 1L, sample_id = NULL)

rowGraph(x, type = 1L, sample_id = NULL)

annotGraph(x, type = 1L, sample_id = NULL)

\S4method{spatialGraph}{SpatialFeatureExperiment,missing,numeric}(x, type, MARGIN, sample_id = NULL) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,numeric,numeric}(x, type, MARGIN, sample_id = NULL) <- value

\S4method{spatialGraph}{SpatialFeatureExperiment,character,numeric}(x, type, MARGIN, sample_id = NULL) <- value

colGraph(x, type = 1L, sample_id = NULL) <- value

rowGraph(x, type = 1L, sample_id = NULL) <- value

annotGraph(x, type = 1L, sample_id = NULL) <- value
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object.}

\item{MARGIN}{As in \code{\link{apply}}. 1 stands for rows and 2 stands for
columns. In addition, 3 stands for spatial neighborhood graphs that
correspond to \code{annotGeometries}.}

\item{sample_id}{Name of the sample the graph is associated with. This is
useful when multiple pieces of tissues are in the same SFE object (say for
a joint dimension reduction and clustering) and the spatial neighborhood is
only meaningful within the same piece of tissue. See the \code{sample_id}
argument in \code{\link{SpatialExperiment}}.}

\item{value}{A \code{listw} object (\code{*Graph}), or a named list of list
of \code{listw} objects (\code{*Graphs}) where the names of the top level
list are \code{sample_id}s when adding graphs for all samples in the margin
of interest, or a list of \code{listw} objects when adding graphs for one
sample in one margin.}

\item{type}{An integer specifying the index or string specifying the name of
the *Graph to query or replace. If missing, then the first item in the
*Graph will be returned or replaced.}
}
\description{
Spatial neighborhood graphs as \code{spdep}'s \code{listw} objects are stored
in the \code{int_metadata} of the SFE object. The \code{listw} class is used
because \code{spdep} has many useful methods that rely on the neighborhood
graph as \code{listw}.
}
