% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{SpatRasterImage}
\alias{SpatRasterImage}
\alias{SpatRasterImage-class}
\alias{show,SpatRasterImage-method}
\title{SpatRaster representation of images in SFE objects}
\usage{
SpatRasterImage(img)

\S4method{show}{SpatRasterImage}(object)
}
\arguments{
\item{img}{A \code{\link{SpatRaster}} or \code{PackedSpatRaster} object.}

\item{object}{A \code{SpatRasterImage} object.}
}
\value{
A \code{SpatRasterImage} object.
}
\description{
\code{SpatialFeatureExperiment} and the \code{Voyager} package work with
images differently from \code{SpatialExperiment}. In SFE and
\code{Voyager}'s, plotting functions for SFE objects, the images can be read
with \code{\link{rast}} and represented as \code{SpatRaster}, so the image is
not entirely loaded into memory unless necessary. Plotting will not load a
large image into memory; rather the image will be downsampled and the
downsampled version is plotted. A \code{SpatRasterImage} object (as of Bioc
3.19 or SFE version 1.6 and above) is a \code{SpatRaster} object but also
inheriting from \code{VirtualSpatialImage} as required by
\code{SpatialExperiment}.
}
\examples{
# Example code
}
\concept{Image classes}
