% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{translateImg}
\alias{translateImg}
\alias{translateImg,BioFormatsImage-method}
\alias{translateImg,EBImage-method}
\title{Translate/shift image in space}
\usage{
\S4method{translateImg}{SpatRasterImage}(x, v)

\S4method{translateImg}{BioFormatsImage}(x, v)

\S4method{translateImg}{EBImage}(x, v)
}
\arguments{
\item{x}{An object of class \code{*Image} as implemented in this package.}

\item{v}{Numeric vector of length 2 to shift the image in the x-y plane.}
}
\value{
A \code{*Image} object of the same class that has been shifted in
  space.
}
\description{
This function shifts the spatial extent of the image in the x-y plane.
}
