% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{transposeImg}
\alias{transposeImg}
\alias{transposeImg,SpatRasterImage-method}
\alias{transposeImg,BioFormatsImage-method}
\alias{transposeImg,EBImage-method}
\title{Transpose images}
\usage{
\S4method{transposeImg}{SpatRasterImage}(x, ...)

\S4method{transposeImg}{BioFormatsImage}(x, resolution = 4L)

\S4method{transposeImg}{EBImage}(x, ...)
}
\arguments{
\item{x}{An object of class \code{*Image} as implemented in this package.}

\item{...}{Ignored. It's there so different methods can all be passed to the
same \code{lapply} in the method for SFE objects.}

\item{resolution}{Resolution to read in from OME-TIFF, defaults to 4, which
is a medium resolution in Xenium.}
}
\value{
For \code{SpatRasterImage} and \code{EBImage}, object of the same
  class. For \code{BioFormatsImage}, the image of the specified resolution is
  read into memory and then the \code{EBImage} method is called, returning
  \code{EBImage}. For the extent: xmin and xmax are switched with ymin and
  ymax.
}
\description{
Swap rows and columns of images. In effect, this will flip the image around
the diagonal running from top left to bottom right.
}
\seealso{
Other image methods: 
\code{\link{SFE-image}},
\code{\link{cropImg}()},
\code{\link{imgRaster}()},
\code{\link{imgSource}()},
\code{\link{mirrorImg}()},
\code{\link{rotateImg}()}
}
\concept{Image and raster}
\concept{image methods}
